refactor backend sockets #244
